move partition and writable to data dir
