+++ src/main/java/org/bgi/flexlab/gaea/data/mapreduce/writable/PairWritable.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>
  *******************************************************************************/
-package org.bgi.flexlab.gaea.tools.mapreduce.bamsort;
+package org.bgi.flexlab.gaea.data.mapreduce.writable;
 
 import org.apache.hadoop.io.WritableComparable;
