+++ src/main/java/org/bgi/flexlab/gaea/tools/mapreduce/annotator/Annotator.java
@@ -27,14 +27,13 @@ import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
 import org.apache.hadoop.mapreduce.lib.input.TextInputFormat;
 import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
 import org.apache.hadoop.mapreduce.lib.output.LazyOutputFormat;
-import org.apache.hadoop.mapreduce.lib.output.MultipleOutputs;
 import org.apache.hadoop.mapreduce.lib.output.TextOutputFormat;
-import org.bgi.flexlab.gaea.data.mapreduce.output.vcf.VCFHdfsWriter;
+import org.bgi.flexlab.gaea.data.mapreduce.partitioner.FirstPartitioner;
+import org.bgi.flexlab.gaea.data.mapreduce.writable.PairWritable;
 import org.bgi.flexlab.gaea.data.structure.header.SingleVCFHeader;
 import org.bgi.flexlab.gaea.framework.tools.mapreduce.BioJob;
 import org.bgi.flexlab.gaea.framework.tools.mapreduce.ToolsRunner;
 import org.seqdoop.hadoop_bam.VCFOutputFormat;
-import org.seqdoop.hadoop_bam.VariantContextWritable;
 
 import java.io.IOException;
 import java.util.ArrayList;
